Use bfloat16 (float32 leading to OOM)

+++ tests/spec_decode/e2e/test_mlp_correctness.py
@@ -33,7 +33,7 @@ SPEC_MODEL = "ibm-granite/granite-3b-code-instruct-accelerator"
 MAX_SPEC_TOKENS = 5
 
 # precision
-PRECISION = "float32"
+PRECISION = "bfloat16"
 
 
 @pytest.mark.parametrize(
